fix: address style issues in outlook email client
Email clients often replace newlines with `<br>` and this created a lot of whitespace. All HTML is on a single line now so that is no longer a problem.
The UTF-8 character in the link behind the header was not rendering correctly, so the full dashboard title serves as a link now.
The page-level CSS was being ignored and in some cases was even showing in the email body. So that was removed completely. We may need to add styles later but these will have to be inline-styles.